pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kompler"
version = "0.1.0"
description = "AI document intelligence platform — make your business documents work for you"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"

dependencies = [
    # API
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "python-multipart>=0.0.9",
    "sse-starlette>=2.0",

    # AI
    "anthropic>=0.40",
    "langgraph>=0.3",
    "langchain-core>=0.3",
    "sentence-transformers>=3.0",

    # Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pgvector>=0.3",

    # Storage
    "boto3>=1.34",

    # Cache & Queue
    "redis>=5.0",
    "arq>=0.26",

    # Data
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.27",

    # Observability
    "structlog>=24.0",
    "sentry-sdk[fastapi]>=2.0",

    # Security & Rate Limiting
    "slowapi>=0.1.9",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",

    # Payments
    "stripe>=8.0",

    # Utils
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "respx>=0.21",
    "ruff>=0.4",
    "mypy>=1.10",
    "httpx>=0.27",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
